Fixed #28144 -- Added allow_overwrite kwarg to FileSystemStorage._save
Inspiration here was taken from django.core.files.move.file_move_safe
which also accepts a boolean `allow_overwrite` kwarg to control the
usage of the O_EXCL flag.

The intention here is to enable Django users to create an "overwriting"
storage backend while still leveraging FileSystemStorage. No change in
behaviour takes place here unless a user passes in allow_override=True
to FileSystemStorage._save. (See recently added test case
FileStorageTests.test_save_overwrite_behaviour)

--

A piece of advice I've seen in several places [1] for creating an
overwriting storage backend is to create a subclass of FileSystemStorage
where the get_available_name deletes an existing file of the same name.
This, in my opinion, is a sneaky use of this function, based on the
name.

A less sneaky (but more race-conditiony) implementation I have seen is to
have _save delete the existing file before calling FileSystemStorage._save

With this commit in place a cleaner implementation of an overwriting
filesystem-backed storage would be possible -- see the
OverwritingStorage (and OverwritingStorageTests) in file_storage tests.

--

[1] Various incarnations of this "ovewriting" implementation:
    - django-storages, marked as deprecated (https://github.com/jschneier/django-storages/blob/1.5.2/storages/backends/overwrite.py)
    - djangosnippets (https://djangosnippets.org/snippets/976/)
    - someone's gist (https://gist.github.com/fabiomontefuscolo/1584462)
    - django-overwrite-storage (https://github.com/ckot/django-overwrite-storage/blob/master/overwrite_storage/storage.py)